[Refactor] post 관련 api 연동 (#2)
* feat: post, cursor 모델 생성

* feat: getPosts 서비스 코드 작성

* feat: post provider 추가

* refactor: post 목록 위젯 수정

* refactor: postBubble 형식에 맞게 model 수정

* feat: 게시글 추가/수정/삭제 service 추가

* refactor: 게시글 생성 로직 수정

* fix: id String에서 int로 변경

* fix: 오타 수정

+++ lib/screens/mypage/screen_myComment.dart
@@ -77,8 +77,8 @@ class MyComment extends StatelessWidget {
                   return PostBubble(
                     userDetail.photoUrl,
                     userDetail.nickname,
-                    boardId,
-                    board.userId,
+                    boardId as int,
+                    board.userId as int,
                     board.body,
                     board.likeCnt,
                     board.commentCnt,
